requirements.txt
feedparser
requests
schedule
yfinance
pandas
google-generativeai
duckduckgo-search